board_be/requirements.txt
Django
psycopg
djangorestframework
djangorestframework-simplejwt
drf-spectacular
uuid
djoser
drf-yasg